chore: split up driver components into separate crates
