+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "hir-transform",
   "hir-type",
   "midenc",
+  "midenc-compile",
   "midenc-driver",
   "midenc-session",
   "tools/*",
@@ -38,6 +39,7 @@ clap = { version = "4.1", features = ["derive", "env"] }
 cranelift-entity = "0.100"
 cranelift-bforest = "0.100"
 env_logger = "0.9"
+either = "1.9"
 Inflector = "0.11"
 intrusive-collections = "0.9"
 inventory = "0.3"
@@ -66,6 +68,7 @@ miden-hir-transform = { path = "hir-transform" }
 miden-hir-type = { path = "hir-type" }
 miden-parsing = "0.1"
 miden-frontend-wasm = { path = "frontend-wasm" }
+midenc-compile = { path = "midenc-compile" }
 midenc-driver = { path = "midenc-driver" }
 midenc-session = { path = "midenc-session" }
