+++ midenc-compile/src/stages/parse.rs
@@ -25,7 +25,7 @@ impl Stage for ParseStage {
         input: Self::Input,
         _analyses: &mut AnalysisManager,
         session: &Session,
-    ) -> DriverResult<Self::Output> {
+    ) -> CompilerResult<Self::Output> {
         use midenc_session::{FileType, InputType};
         // Track when compilation began
         let file_type = input.file_type();
@@ -44,7 +44,7 @@ impl Stage for ParseStage {
     }
 }
 impl ParseStage {
-    fn parse_ast_from_file(&self, path: &Path, session: &Session) -> DriverResult<ParseOutput> {
+    fn parse_ast_from_file(&self, path: &Path, session: &Session) -> CompilerResult<ParseOutput> {
         use std::io::Read;
 
         let mut file = std::fs::File::open(path)?;
@@ -53,12 +53,12 @@ impl ParseStage {
         self.parse_ast_from_bytes(&bytes, session)
     }
 
-    fn parse_ast_from_bytes(&self, bytes: &[u8], session: &Session) -> DriverResult<ParseOutput> {
+    fn parse_ast_from_bytes(&self, bytes: &[u8], session: &Session) -> CompilerResult<ParseOutput> {
         use miden_hir::parser::Parser;
         use std::io::{Error, ErrorKind};
 
         let source = core::str::from_utf8(bytes).map_err(|_| {
-            DriverError::Io(Error::new(
+            CompilerError::Io(Error::new(
                 ErrorKind::InvalidInput,
                 "input is not valid utf-8",
             ))
@@ -74,7 +74,7 @@ impl ParseStage {
         &self,
         path: &Path,
         session: &Session,
-    ) -> DriverResult<ParseOutput> {
+    ) -> CompilerResult<ParseOutput> {
         use std::io::Read;
 
         let mut file = std::fs::File::open(path)?;
@@ -87,7 +87,7 @@ impl ParseStage {
         &self,
         bytes: &[u8],
         session: &Session,
-    ) -> DriverResult<ParseOutput> {
+    ) -> CompilerResult<ParseOutput> {
         let config = wasm::WasmTranslationConfig::default();
         let module = wasm::translate_module(bytes, &config, &session.diagnostics)?;
