--- midenc-driver/src/commands/mod.rs
@@ -1,5 +1,0 @@
-mod compile;
-mod debug;
-
-pub use self::compile::compile;
-pub use self::debug::Breakpoint;
